+++ pkg/util/validation/validate.go
@@ -65,6 +65,15 @@ func metricReasonFromErrorID(id globalerror.ID) string {
 	return strings.ReplaceAll(string(id), "-", "_")
 }
 
+// DiscardedRequests is a metric of the number of discarded requests.
+var DiscardedRequests = prometheus.NewCounterVec(
+	prometheus.CounterOpts{
+		Name: "cortex_discarded_requests_total",
+		Help: "The total number of requests that were discarded due to rate limiting.",
+	},
+	[]string{discardReasonLabel, "user"},
+)
+
 // DiscardedSamples is a metric of the number of discarded samples, by reason.
 var DiscardedSamples = prometheus.NewCounterVec(
 	prometheus.CounterOpts{
@@ -269,6 +278,9 @@ func ValidateMetadata(cfg MetadataValidationConfig, userID string, metadata *mim
 func DeletePerUserValidationMetrics(userID string, log log.Logger) {
 	filter := map[string]string{"user": userID}
 
+	if err := util.DeleteMatchingLabels(DiscardedRequests, filter); err != nil {
+		level.Warn(log).Log("msg", "failed to remove cortex_discarded_requests_total metric for user", "user", userID, "err", err)
+	}
 	if err := util.DeleteMatchingLabels(DiscardedSamples, filter); err != nil {
 		level.Warn(log).Log("msg", "failed to remove cortex_discarded_samples_total metric for user", "user", userID, "err", err)
 	}